[DEV-4375] Split account models
